requirements.txt
pillow
selenium
pyscreenshot